feat: add real-time SVG preview window to UI
- Extracted spiral generation logic into `src/spiral.ts`
- Added square preview window rendering SVG path instantly without crossing Figma IPC
- Set viewBox to center on (0,0) and scale to bounding box symmetrically
- Added percentage scale display in bottom right
- Renamed Continuously Update button text
